+++ database/migrations/2020_12_10_042436_create_kontrols_table.php
@@ -18,10 +18,10 @@ class CreateKontrolsTable extends Migration
             $table->integer('order');
             $table->date('date');
             $table->string('tempat_kontrol');
-            $table->integer('weight')->nullable();
-            $table->integer('length')->nullable();
-            $table->integer('lingkar_kepala')->nullable();
-            $table->integer('temperature');
+            $table->float('weight')->nullable();
+            $table->float('length')->nullable();
+            $table->float('lingkar_kepala')->nullable();
+            $table->float('temperature');
             $table->foreignId('patient_profile_id');
             $table->string('note', 500)->nullable();
             $table->string('nurse_note', 500)->nullable();
